requirements.txt
asgiref==3.6.0
Django==4.1.7
django-countries==7.5.1
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
djangorestframework==3.14.0
phonenumbers==8.13.7
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.5.0
